Remove dynamic field names from the packages index
So far, dynamic field names have been used for useflags in the
packages index. This leads to a large number of fields, which
is continuously growing, as obsolete fields are not cleaned up
from the mapping when the useflag is deleted. This way, the
maximum limit of 50000 fields may be hit.

To avoid this, the dynamic field names have been removed
in favour of a simple list. This way it was possible to reduce
the number of fields from nearly 50000 to 41.

Please note:
Due to the change of the mapping, the new index will no longer
be compatible with the old one. A fresh index has to be populated.

Signed-off-by: Max Magorsch <max@magorsch.de>