Add further mappings for the indexes
The mapping for maintainers field in the package index has been added.
Furthermore, the commit field in the change index has been changed from
object to keyword. However, currently, it is not used. It may be removed
in future versions. Finally, missing id fields have been added to the
mappings.

Signed-off-by: Max Magorsch <max@magorsch.de>