Turn off dynamic mapping in all indexes
By default, fields are added dynamically, whenever they are not present
in the current mapping. However, we exactly specify the mapping, so we
don't want any unexpected fields to be added to the index.

That's why we use 'strict' here. This way we will get an exception every
time an unexpected field is detected (which should not happen).

Signed-off-by: Max Magorsch <max@magorsch.de>