Fix the broken feeds
Due to the migration to the repository pattern in
elasticsearch-persistence the fields created_at and
updated_at are no longer automatically parsed. That's
why the feeds have been broken, as they rely on the
parsed dates.
To fix this problem, we manually parse the dates during
the deserialization now.

Fixes https://bugs.gentoo.org/697666

Signed-off-by: Max Magorsch <max@magorsch.de>