Add a query parser for advanced search queries
Add a query parser using parslet which is able to parse advanced
queries. This way, it is possible to run queries like:
  > +maintainers.emails:haskell@gentoo.org
or
  > +maintainers.emails:haskell@gentoo.org -category:dev-haskell
to find all packages that are maintained by the haskell herd and
respectively all packages that are maintained by the haskell herd
but not part of dev-haskell.
Documentation on how to use the advanced queries is included
into the website and can be viewed by clicking on a question mark
next to the search bar on the landing page.

Signed-off-by: Max Magorsch <max@magorsch.de>