Add atom feeds for package search results
Add atom feeds for the results of search queries. For instance you
can subscribe to the atom feed for the result of the query
  > +maintainers.email:haskell@gentoo.org
now. This way you are able to keep track of packages that belong to
a certain herd/maintainer.
The atom feed contains the latest commit for each package so far.

Bug: https://bugs.gentoo.org/408661

Signed-off-by: Max Magorsch <max@magorsch.de>