Change the documentation link to the wiki page
So far the documentation link for each package points to a full-text
search in the wiki. Now the documentation link points to the respective
wiki page in case the page exists. Only in case, the page doesn't exist
the link points to the full-text search.
Additionally, if the page doesn't exist the link is labelled as
'Search for Documentation'. In case the page exists, it is still
labelled 'Documentation'.

Closes: https://bugs.gentoo.org/611262

Signed-off-by: Max Magorsch <max@magorsch.de>