Migrate the project from sprockets to webpacker
The asset pipeline was introduced in Rails 3.1. However, since Rails 5.1
webpacker has been available, so that it's possible to use webpack.

The project has been fully migrated to use webpacker for bundeling
javascripts as well as stylesheets now. This way, sprockets has been
completely replaced and removed from the project. Associated gems as
jquery-rails have been removed as well.
Accordingly all advanced webpack functionalities are available now.

The bin/first-run file as well as the Dockerfiles have been adjusted to
use webpacker instead of the asset pipeline.

Please note: In order to use webpacker, yarn has to be installed on
the target system. Please make sure that 'yarnpkg' is in your path.

Signed-off-by: Max Magorsch <max@magorsch.de>