Update to Bootstrap 4
Update the project to Bootstrap 4, which is a major rewrite of the
entire project. So far, a Tyrian version that is based on Bootstrap 4
is included for that reason. However, in future, it may be retrieved
from the CDN or installed as node module.

Webpacker is bundeling all of the javascript code and stylesheets.

Signed-off-by: Max Magorsch <max@magorsch.de>