Fix the useflag search for USE flags with hyphens
So far, 'name' was mapped as a 'text'-field in the elasticsearch
useflag index. Accordingly, values are analyzed when entering them
into elasticsearch. This is favourable, when suggesting useflags for
typeahead for instance, as match_phrase_prefix can be used. However,
when searching for a specific useflag by name this leads to problems,
as for instance searching for gmxapi-legacy returns:
  - gmxapi-legacy
  - gmxapi
  - legacy
as the field is analyzed.

That's why an additional raw field has been added to the 'name' field
using multi-fields. This way name and name.raw can be used as follows:
  - name is of type text
  - name.raw is of type keyword
Accordingly name.raw can be used when searching for a specific useflag
by name, while it is still possible to use name for match_phrase_prefix
and typeahead.

Please see https://www.elastic.co/guide/en/elasticsearch/reference/
current/multi-fields.html for more details.

PLEASE NOTE: As the index has been changed, a new index has to
populated after this commit.

Signed-off-by: Max Magorsch <max@magorsch.de>